Implement Core-Platform communication through App
This branch is an attempt to establish reliable and efficient
communication of Core and Platform classes through App.

The one major benefit of this would be clear separation of concerns:

Platform is designed to be a socket for any particular OS, sending
OS-agnostic events to Core.

Core is the program functionality implementation. It must not be
concerned with the OS or Platform.

App is the in-between layer, transferring messages between the two. It is
also going to load and hold user config, among other global-state data.

The communication is built as callback functions platformToCore and
coreToPlatform.