Implement BasePlatform class for holding key codes
Previously BasePlatform was deleted in order to simplify the
architecture. However, upon closer examination, it appears to be more
sensible to have BasePlatform class handle common data fields, and leave
IPlatform class, which is inherited by BasePlatform, as an interface.

BasePlatform encapsulates two std::unordered_maps that handle
Key-NativeKeyCode conversions when sending events from Core to Platform
and vice versa.

All funciton members of BasePlatform (not inherited from IPlatform) are
static for some of them are to be called before any object of the class
is instantiated, and all of them only need access to static (as they are
shared by all instances) data members.