Implement Core::getRemappedKeys()
Currently, when HRMMode is enabled, the event is sent from Platform to
Core, where it is remapped and sent back. This back and forth
communication introduces considerable latency to key presses. It is to
be considered to remap keys in place, and modify incoming events instead
of nullifying it, and generating another.