Separate modifiers from printable keys
The separation is needed to distinguish events coming from the Platform,
from those comming from the App.

The Platform only sends printable characters, which are then converted
by the App to the user-binded modifiers and stored in a temporary
object.

Once a non-binded key arrives from the Platform, the App sends then
temporary object back, so the modifiers can be added to the non-binded
key event, and sends it back to the OS.