Implement entering HRM mode on a leader key press
Since only a modifier key can be the leader, and pressing and releasing
it introduces two separate events on MacOS, for now it was decided to
keep track of whether or not the release of the leader key was
processed. It is not the cleanest design possible, and should be
revisited.