Fix non-consecutive combinations
Previously the combination wasn't getting registered if the leader and a
binded key were pressed at the same time. This is now fixed by checking
the keycode when processing event of the leader key going up. Before the
changes in this commit were made, any key press that was made before the
leader key went up was being registered as it.